add median filter test
